remove local tunefile. change inj map type again, some issue in previous commit

+++ asf-efi.ino
@@ -19,7 +19,7 @@ volatile bool     pump_active = false;
 bool              pump_manual = false;
 bool              pump_mode_always_on = false;
 
-uint16_t inj_map[RPM_BINS][TPS_BINS];
+uint8_t inj_map[RPM_BINS][TPS_BINS];
 uint16_t iat_correction[IAT_CORR_BINS];
 uint16_t et_correction[ET_CORR_BINS];
